+++ src/Proxy/PolygonProxy.php
@@ -316,4 +316,13 @@ class PolygonProxy extends Polygon implements ProxyInterface
         return $this->proxyGeometry->swapXY();
     }
 
+    public function isIdenticalTo(\Brick\Geo\Geometry $that) : bool
+    {
+        if ($this->proxyGeometry === null) {
+            $this->load();
+        }
+
+        return $this->proxyGeometry->isIdenticalTo($that);
+    }
+
 }