+++ src/Proxy/PolyhedralSurfaceProxy.php
@@ -298,4 +298,13 @@ class PolyhedralSurfaceProxy extends PolyhedralSurface implements ProxyInterface
         return $this->proxyGeometry->swapXY();
     }
 
+    public function isIdenticalTo(\Brick\Geo\Geometry $that) : bool
+    {
+        if ($this->proxyGeometry === null) {
+            $this->load();
+        }
+
+        return $this->proxyGeometry->isIdenticalTo($that);
+    }
+
 }