+++ src/Proxy/SurfaceProxy.php
@@ -271,6 +271,15 @@ class SurfaceProxy extends Surface implements ProxyInterface
         return $this->proxyGeometry->project($projector);
     }
 
+    public function isIdenticalTo(\Brick\Geo\Geometry $that) : bool
+    {
+        if ($this->proxyGeometry === null) {
+            $this->load();
+        }
+
+        return $this->proxyGeometry->isIdenticalTo($that);
+    }
+
     public function count()
     {
         if ($this->proxyGeometry === null) {