+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 - New `Projector` API to reproject geometry coordinates.
 - New method: `Polygon::rings()`
+- New method: `Geometry::isIdenticalTo()`
 
 ## [0.8.1](https://github.com/brick/geo/releases/tag/0.8.1) - 2022-10-14